pyproject.toml
[tool.poetry]
name = "truthness"
version = "0.1.0"
description = ""
authors = ["seeusoon <to.asmarkov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
python-box = "^7.2.0"
langchain-core = "^0.3.16"
langchain-openai = "^0.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
